添加loading样式
