+++ package.json
@@ -30,6 +30,7 @@
     "react": "16.3.2",
     "react-dom": "16.3.2",
     "react-icons": "^2.2.7",
+    "react-loadable": "^5.4.0",
     "react-redux": "^5.0.6",
     "react-router-dom": "^4.2.2",
     "react-transition-group": "2.2.1",
