+++ public/index.html
@@ -37,4 +37,5 @@
       To create a production bundle, use `npm run build` or `yarn build`.
     -->
   </body>
+  <script src="https://cdn.polyfill.io/v2/polyfill.min.js?features=Map,Set"></script>
 </html>
