+++ src/containers/ArchiveArticles/index.js
@@ -1,103 +1,17 @@
 import React from 'react';
-import {connect} from 'react-redux';
-import {withRouter} from 'react-router-dom';
-import NProgress from 'nprogress';
-import Card from '../../components/Card';
-import '../Home/home.scss';
-import {
-    findMatchTagsArticle,
-    reduceVisit,
-    getPartArticles,
-} from "../../reducers/article.redux";
-import ReadMore from '../../components/ReadMore';
-import BottomOut from "../../components/BottomOut";
-import RightSideBar from "../RightSideBar";
+import Loadable from 'react-loadable';
+import Loading from '../../components/Loading';
 
-class ArchiveArticles extends React.PureComponent {
+const LoadableComponent = Loadable({
+    loader: () => import('./LoadableArchiveArticles'),
+    loading: Loading,
+});
 
-    constructor(props) {
-        super(props);
-        this.state = {
-            skip: 0,
-            limit: 6
-        };
-    }
-
-    showPostContent(articleId, visit) {
-        this.props.history.push({
-            pathname: `/article/${articleId}`
-        });
-        this.props.reduceVisit({id: articleId, visit: visit + 1});
-    }
-
-    componentDidMount() {
-        const {tagName}  = this.props.match.params;
-        NProgress.start();
-        this.props.findMatchTagsArticle({tag: [tagName]});
-    }
-
-    componentDidUpdate() {
-        NProgress.done();
-    }
-
-    componentWillUnmount() {
-        NProgress.done();
-    }
-
-    tagClick(v) {
-        this.props.history.push(`/tag/${v}`);
-        this.props.findMatchTagsArticle({tag: v});
-    }
-
-    renderCards(v, index) {
-        return <Card key={index} articleId={v.id} title={v.title} thumb={v.thumb} visit={v.visit}
-                     summary={v.summary} tags={v.tags} date={v.date} clickTag={(v) => this.tagClick(v)}
-                     showPost={(id) => this.showPostContent(id, v.visit)} showCardInfo={true}/>;
-    }
-
-    readMore(v) {
-        this.setState(state => ({
-            limit: state.limit + 3
-        }));
-    }
-
-    renderReadMore(filled) {
-        return filled ? <BottomOut/> : <ReadMore handleReadMore={(v) => this.readMore(v)}/>;
-    }
-
-    renderPartArticle(partArticle) {
-        return partArticle ?
-            partArticle.filter(v => v.publish).map((v, index) => (
-                this.renderCards(v, index)
-            )) :'';
-    }
-
-    render() {
-        const {article} = this.props;
-        const pageSize = article.length;
-        const partArticle = article.slice(0, this.state.limit);
+export default class ArchiveArticles extends React.Component {
+    render () {
         return (
-            <div className='container'>
-                <div className={'articles'}>
-                    { this.renderPartArticle(partArticle)  }
-                    {  this.renderReadMore(this.state.limit >= pageSize) }
-                </div>
-                <RightSideBar showPostContent={(id, visit) => this.showPostContent(id, visit)}
-                              articleSideBarTitle={'热门文章'} showPopular={true}
-                />
-            </div>
+            <LoadableComponent/>
         );
     }
 }
 
-const mapStateToProps = state => {
-    return {
-        article: state.matchTagArticles,
-    };
-};
-
-export default withRouter(connect(mapStateToProps, {
-    getPartArticles,
-    reduceVisit,
-    findMatchTagsArticle,
-})(ArchiveArticles));
